feat(app-info): show the tabbed details in the sheet itself
AppInfoDialog becomes AppInfoSheet: one surface instead of two. The sheet
opens at the partial detent with the header and the action row, and dragging
it to Expanded reveals the same AppInfoDetailBody the details screen uses,
in place. Nothing navigates away any more, so AppClickAction.OpenDetails
loses its only producer and goes with it, along with the freezer's
never-reachable AppInfoDetails landing.

The content column fills the sheet height and scrolls, and the detail body is
sized against the sheet's own height rather than against what the header
leaves over. A height-bounded Column squeezes its overflow instead of
clipping it, so on a short window (landscape, split screen, a large font
scale) a weighted body would measure 0dp and take the action labels with it.
Filling also keeps the anchors still: the sheet's measured height no longer
depends on whether the body has arrived.

The "Details" action does both halves of the drag. expand() only moves the
anchor and is a plain no-op once the sheet is already Expanded, which on a
short window is the only state the action row is reachable from - on its own
it would be a control that is guaranteed to do nothing there. It now settles
the sheet and then scrolls the body up, matching what the nested-scroll
connection does for a real drag.

Detail parsing is scoped to the composable via rememberViewModelStoreOwner
and gated on the sheet heading for Expanded, so opening the sheet to tap
Force-stop costs nothing and closing it releases the parse.

Both hosts keep the selected package in rememberSaveable and resolve it
against the unfiltered scan, never the filtered list: rememberSaveable
outlives the process while appListType does not, so resolving against
displayedApps would restore a system-app selection that nothing can render
and nothing can clear, then spring the sheet open on a later tap of the
System toggle.
