+++ app/src/main/java/com/valhalla/thor/domain/model/AppClickAction.kt
@@ -20,6 +20,5 @@ sealed interface AppClickAction {
     data class Suspend(val appInfo: AppInfo) : AppClickAction
     data class UnSuspend(val appInfo: AppInfo) : AppClickAction
     data class ManagePermissions(val appInfo: AppInfo) : AppClickAction
-    data class OpenDetails(val appInfo: AppInfo) : AppClickAction
     data class AddToHomeScreen(val appInfo: AppInfo) : AppClickAction
 }
